requirements.txt
flask
PyPDF2
python-dotenv
google-generativeai